backend/requirements.txt
werkzeug==2.0.3
flask==2.0.3
azure-ai-textanalytics==5.2.0
python-dotenv==0.19.0
flask-cors==3.0.10

pytest==7.3.1
pytest-cov==4.1.0
requests-mock==1.11.0